+++ Changes.md
@@ -4,6 +4,7 @@ Changelog: Faktory || [Faktory Pro](https://github.com/contribsys/faktory/blob/m
 
 ## HEAD
 
+- Fix crash with invalid mutate API usage [#313]
 - Fix form handler on Scheduled Job and Retry Job UI pages [#236]
 - Refactor scheduled job processing to handle large job counts [#309]